% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIFAPAtxt.R
\name{downloadIFAPAtxt}
\alias{downloadIFAPAtxt}
\title{Climatic data from "Instituto de Investigacion y Formacion Agraria y Pesquera" ("Institute of Agricultural Research and Training and Fisheries") of Andalusia}
\usage{
downloadIFAPAtxt(file, c_province, c_station, start = "31-01-2000",
  end = "31-12-2020")
}
\arguments{
\item{file}{the name of the file for the download}

\item{c_province}{string with the code of the province}

\item{c_station}{string with the code of the station}

\item{start}{start date for download}

\item{end}{end date for download data}
}
\value{
A data file as it is from unzip the download file from server
}
\description{
Function to download raw txt file with data from one station
}
\examples{
 \dontrun{
 downloadIFAPAtxt(file = 'IFAPA.txt', c_province = '11', c_station = '5',
                  start = "01-01-2010", end = '31-01-2010')
 }
}
\author{
Manuel Munoz-Marquez \email{manuel.munoz.marquez@gmail.com}
}

