% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIFAPAzip.R
\name{downloadIFAPAzip}
\alias{downloadIFAPAzip}
\title{Climatic data from "Instituto de Investigacion y Formacion Agraria y Pesquera" ("Institute of Agricultural Research and Training and Fisheries") of Andalusia}
\usage{
downloadIFAPAzip(file, c_province, c_station, start = "31-01-2000",
  end = "31-12-2020")
}
\arguments{
\item{file}{the name of the file for the download}

\item{c_province}{string with the code of the province}

\item{c_station}{string with the code of the station}

\item{start}{start date for download}

\item{end}{end date for download data}
}
\value{
A zip file as it is download from server
}
\description{
Function to download raw zip file with data from one station
}
\examples{
 \dontrun{
downloadIFAPAzip(file = 'IFAPA.zip', c_province = '11', c_station = '5',
                 start = "01-01-2010", end = '31-01-2010')
}
}
\author{
Manuel Munoz-Marquez \email{manuel.munoz.marquez@gmail.com}
}

