% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaData.R
\name{getMetaData}
\alias{getMetaData}
\title{Get metadata from "Estaciones Agroclimaticas de Andalucia"}
\usage{
getMetaData(provincia = NA, estacion = NA, nombre_estacion = NA)
}
\arguments{
\item{provincia}{numerial code of the target province}

\item{estacion}{numerical code of the target meteorogical station.}

\item{nombre_estacion}{name of the meteorological station (optional). Only used as label.}
}
\value{
onw row data.frame with metadata of the station
}
\description{
This function retrives the metadata of the targeted meteorological station.

See \url{https://www.juntadeandalucia.es/agriculturaypesca/ifapa/ria/servlet/FrontController?action=Init} for explanation of the
data and codes.
}
\examples{
\dontrun{
provincia <- c(41,11)
estacion <- c(5,1)
nombre_estacion <- c("Aznalcollar", "Basulta")
getMetaData(provincia, estacion, nombre_estacion)
getOneMetaData(41, 5, "aznalcollar")
}
}
\author{
Ignasi Bartomeus \email{nacho.bartomeus@gmail.com}
}

