% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readallIFAPA.R
\name{readallIFAPA}
\alias{readallIFAPA}
\title{Climatic data from "Instituto de Investigacion y Formacion Agraria y Pesquera"
 ("Institute of Agricultural Research and Training and Fisheries") of Andalusia}
\usage{
readallIFAPA(start = "31-01-2000", end = "31-12-2020")
}
\arguments{
\item{start}{start date for download}

\item{end}{end date for download data}
}
\value{
A data.frame with the data download after some transformations
}
\description{
Function to download and read data from all stations
}
\examples{
 \dontrun{

readallIFAPA(start = "01-01-2010", end = '31-01-2010')

}

}
\author{
Manuel Munoz-Marquez \email{manuel.munoz.marquez@gmail.com}
}

