% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIFAPA.R
\name{readIFAPA}
\alias{readIFAPA}
\title{Climatic data from "Instituto de Investigacion y Formacion Agraria y Pesquera" ("Institute of Agricultural Research and Training and Fisheries") of Andalusia}
\usage{
readIFAPA(c_province, c_station, start = "31-01-2000", end = "31-12-2020")
}
\arguments{
\item{c_province}{string with the code of the province}

\item{c_station}{string with the code of the station}

\item{start}{start date for download}

\item{end}{end date for download data}
}
\value{
A data.frame with the data download after some transformations
}
\description{
Function to download and read data from one station

## @param file the name of the file for the download
}
\examples{
 \dontrun{
readIFAPA(c_province = '11', c_station = '5', start = "01-01-2010", end = '31-01-2010')
}
}
\author{
Manuel Munoz-Marquez \email{manuel.munoz.marquez@gmail.com}
}

